Big cleanup of StreamArray construction
