--- VisioAutomation_2010/VisioAutomation/ShapeSheet/Streams/StreamBuilder.cs
@@ -1,44 +1,0 @@
-using System.Collections.Generic;
-
-namespace VisioAutomation.ShapeSheet.Streams
-{
-    public abstract class StreamBuilder<T>
-    {
-        protected readonly List<T> _items;
-
-        protected StreamBuilder()
-        {
-            this._items = new List<T>();
-        }
-
-        public int Count => this._items.Count;
-
-        protected StreamBuilder(int capacity)
-        {
-            this._items = new List<T>(capacity);
-        }
-
-        public void Add(T item)
-        {
-            this._items.Add(item);
-        }
-
-        public void AddRange(IEnumerable<T> items)
-        {
-            this._items.AddRange(items);
-        }
-
-        public void Clear()
-        {
-            this._items.Clear();
-        }
-
-        public StreamArray ToStream()
-        {
-            var stream = this.build_stream();
-            return stream;
-        }
-
-        protected abstract StreamArray build_stream();
-    }
-}
