+++ VisioAutomation_2010/VisioAutomation/VisioAutomation.csproj
@@ -113,7 +113,6 @@
     <Compile Include="ShapeSheet\Writers\SidSrcWriter.cs" />
     <Compile Include="ShapeSheet\Writers\SrcWriter.cs" />
     <Compile Include="ShapeSheet\Streams\StreamType.cs" />
-    <Compile Include="ShapeSheet\Streams\StreamBuilderUtil.cs" />
     <Compile Include="ShapeSheet\Writers\WriteCache.cs" />
     <Compile Include="ShapeSheet\Streams\StreamArray.cs" />
     <Compile Include="Shapes\CustomPropertyType.cs" />
@@ -121,7 +120,6 @@
     <Compile Include="ShapeSheet\Internal\ArraySegment.cs" />
     <Compile Include="ShapeSheet\Query\SectionQueryCache.cs" />
     <Compile Include="ShapeSheet\Query\Rows.cs" />
-    <Compile Include="ShapeSheet\Streams\StreamBuilder.cs" />
     <Compile Include="ShapeSheet\Query\Columns.cs" />
     <Compile Include="ShapeSheet\CellGroups\CellGroupBuilder.cs" />
     <Compile Include="ShapeSheet\Query\Column.cs" />
